fixed startup problem
caused by changes for #24.
apparently having a DAT called "modules" inside a "local" Base COMP
causes python errors in mysterious places. it sort of makes sense given
that name is supposed to be used for a Base COMP that contains python
modules, but it's still frustrating that it took me that long to figure
it out.