reimplemented the preset system
It now stores data in parameters on the modules, rather than in module
storage. Presets UI panels are accessible via the module UI mode menu.
closes #270.
It includes a button to reset the whole module to defaults (closes #82).
It includes editable preset names (closes #73).
It still needs to be copied into all the modules.